+++ lib/Screens/LoginAndSignupScreen/LoginAndSignupScreenViewModel.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/foundation.dart';
 import 'package:rxdart/subjects.dart';
+
 import 'package:splitter/Models/User.dart';
 import 'package:splitter/Services/AuthenticationService.dart';
 import 'package:splitter/Services/CloudStoreService.dart';
