+++ lib/Services/AuthenticationService.dart
@@ -1,8 +1,8 @@
 
 import 'package:flutter/foundation.dart';
 import 'package:rxdart/rxdart.dart';
-
 import 'package:firebase_auth/firebase_auth.dart';
+
 import 'package:splitter/Models/AuthenticationState.dart';
 
 abstract class AuthenticationServiceType {
