+++ lib/Services/CloudStoreService.dart
@@ -1,6 +1,8 @@
 import 'dart:async';
+
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/foundation.dart';
+
 import 'package:splitter/Models/User.dart';
 
 abstract class CloudStoreServiceType {
