+++ test/RootScreenViewModel_test.dart
@@ -2,11 +2,14 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:rxdart/rxdart.dart';
+
 import 'package:splitter/Models/AuthenticationState.dart';
 import 'package:splitter/Models/User.dart';
 import 'package:splitter/Screens/RootScreen/RootScreenViewModel.dart';
 import 'package:splitter/Services/AuthenticationService.dart';
 import 'package:splitter/Services/CloudStoreService.dart';
+
+import 'TestStubs.dart';
 
 class FirebaseAuthMock extends Mock implements FirebaseAuth { }
 
@@ -44,65 +47,52 @@ void main() {
       test('ViewModel AuthenticationState is loggedOut when authService emits loggedOut', () async {
         authenticationServiceState.add(AuthenticationState.loggedOut);
 
-        List<AuthenticationState> expectation = [AuthenticationState.loading, AuthenticationState.loggedOut];
+        List<AuthenticationState> expectedStates = [AuthenticationState.loading, AuthenticationState.loggedOut];
 
-        expect(viewModel.authenticationState, emitsInOrder(expectation));
+        expect(viewModel.authenticationState, emitsInOrder(expectedStates));
       });
 
       test('ViewModel AuthenticationState is loggedIn when authService emits loggedIn', () async {
-        String userID = "mockUserID";
         User userStub = TestUser.stub();
-        List<AuthenticationState> expectation = [AuthenticationState.loading, AuthenticationState.loggedIn];
+        List<AuthenticationState> expectedStates = [AuthenticationState.loading, AuthenticationState.loggedIn];
 
         when(authenticationService.currentUserId()).thenAnswer((_) =>
-          Future<String>.value(userID));
-        when(cloudStoreService.fetchUserWithId(userID)).thenAnswer((_) =>
+          Future<String>.value(userStub.id));
+        when(cloudStoreService.fetchUserWithId(userStub.id)).thenAnswer((_) =>
           Future<User>.value(userStub));
 
         authenticationServiceState.add(AuthenticationState.loggedIn);
 
-        expect(viewModel.authenticationState, emitsInOrder(expectation));
+        expect(viewModel.authenticationState, emitsInOrder(expectedStates));
       });
     });
 
     group('On Error', () {
-      test('ViewModel AuthenticationState is loggedOut when currentUserId errors', () async {
-        List<AuthenticationState> expectation = [AuthenticationState.loading, AuthenticationState.loggedOut];
+      test('ViewModel AuthenticationState is loggedOut when currentUserId throws', () async {
+        List<AuthenticationState> expectedStates = [AuthenticationState.loading, AuthenticationState.loggedOut];
 
         when(authenticationService.currentUserId()).thenThrow((_) =>
-          Future<String>.value("error"));
+          Future<String>.value(null));
 
         authenticationServiceState.add(AuthenticationState.loggedIn);
 
-        expect(viewModel.authenticationState, emitsInOrder(expectation));
+        expect(viewModel.authenticationState, emitsInOrder(expectedStates));
       });
 
-      test('ViewModel AuthenticationState is loggedOut when fetchUserWithId errors', () async {
-        User testUser = TestUser.stub();
-        List<AuthenticationState> expectation = [AuthenticationState.loading, AuthenticationState.loggedOut];
+      test('ViewModel AuthenticationState is loggedOut when fetchUserWithId throws', () async {
+        User userStub = TestUser.stub();
+        List<AuthenticationState> expectedStates = [AuthenticationState.loading, AuthenticationState.loggedOut];
 
         when(authenticationService.currentUserId()).thenAnswer((_) =>
-          Future<String>.value(testUser.id));
+          Future<String>.value(userStub.id));
 
-        when(cloudStoreService.fetchUserWithId(testUser.id)).thenThrow((_) =>
-          Future<String>.value("error"));
+        when(cloudStoreService.fetchUserWithId(userStub.id)).thenThrow((_) =>
+          Future<String>.value(null));
 
         authenticationServiceState.add(AuthenticationState.loggedIn);
 
-        expect(viewModel.authenticationState, emitsInOrder(expectation));
+        expect(viewModel.authenticationState, emitsInOrder(expectedStates));
       });
     });
   });
-}
-
-extension TestUser on User {
-    static User stub([String id = "mockId", 
-                      String firstName = "mockFirstName",
-                      String lastName = "mockLastName",
-                      String email = "mockEmail"]) {
-      return User(id: id,
-                  firstName: firstName, 
-                  lastName: lastName, 
-                  email: email);
-    }
 }
